Calculator Assignment

+++ src/main/java/se/lexicon/calculator.java
@@ -6,12 +6,12 @@ public class calculator {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Please enter your first number.");
-        int num1 = scanner.nextInt();
+        double num1 = scanner.nextDouble();
         System.out.println("Please enter your operator.");
         char operator = scanner.next().charAt(0);
         System.out.println("Please enter your second number.");
-        int num2 = scanner.nextInt();
-        int result = 0;
+        double num2 = scanner.nextDouble();
+        double result = 0;
 
         switch (operator) {
             case '+':
